requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
pandas==2.2.3
python-dateutil==2.8.2
pytest==7.4.0
httpx==0.24.1
Faker==18.9.1